File changes and small fix
Removed pointless, outdated zip file and added a separate JS file for the HTML file to refer to, as I would like to keep things more organised (which I'm not that good at) and it lets Github display the true percentage of different code types. I don't about the effects on performance but that doesn't really matter as this is such a low stress project and also idrc. I also fixed the rounding error (not the ugrade) on the score per second display.